+++ bgstally/bgstally.py
@@ -156,8 +156,9 @@ class BGSTally:
                 dirty = True
 
             case 'MissionAccepted':
-                self.mission_log.add_mission(entry.get('Name', ""), entry.get('Faction', ""), entry.get('MissionID', ""), entry.get('Expiry', ""), system, station,
-                    entry.get('Count', -1), entry.get('PassengerCount', -1), entry.get('KillCount', -1))
+                self.mission_log.add_mission(entry.get('Name', ""), entry.get('Faction', ""), entry.get('MissionID', ""), entry.get('Expiry', ""),
+                                             entry.get('DestinationSystem', ""), entry.get('DestinationSettlement', ""), system, station,
+                                             entry.get('Count', -1), entry.get('PassengerCount', -1), entry.get('KillCount', -1))
                 dirty = True
 
             case 'MissionCompleted':